refactor: move updateDynamicImage to controller.js
The updateDynamicImage function is extracted from the dynamic-hero.js
block module and moved into the controller.js script. This change
consolidates image updating logic in the controller, allowing it to be
reused across different components.

The dynamic-hero block now imports the updateDynamicImage function and
calls it after decorating the block, ensuring the hero image is still
updated dynamically.

This refactoring improves code organization and reusability by moving
related functionality into a centralized controller module.
